requirements.txt
streamlit
pandas
numpy
matplotlib.pyplot
altair
plotly_express